+++ bot/__main__.py
@@ -2,8 +2,8 @@ import logging
 
 from bot.bot import bot
 from bot.constants import Client, STAFF_ROLES, WHITELISTED_CHANNELS
-from bot.decorators import in_channel_check
 from bot.seasons import get_extensions
+from bot.utils.decorators import in_channel_check
 
 log = logging.getLogger(__name__)
