+++ bot/seasons/evergreen/game.py
@@ -12,7 +12,7 @@ from discord.ext.commands import Cog, Context, group
 
 from bot.bot import SeasonalBot
 from bot.constants import STAFF_ROLES, Tokens
-from bot.decorators import with_role
+from bot.utils.decorators import with_role
 from bot.utils.pagination import ImagePaginator, LinePaginator
 
 # Base URL of IGDB API
