+++ bot/seasons/halloween/candy_collection.py
@@ -9,7 +9,7 @@ import discord
 from discord.ext import commands
 
 from bot.constants import Channels, Month
-from bot.decorators import in_month
+from bot.utils.decorators import in_month
 
 log = logging.getLogger(__name__)
